Rename PromQL transpiler package from "main"->"promflux"
