+++ promflux/query_helpers_test.go
@@ -1,4 +1,4 @@
-package main
+package promflux
 
 import (
 	"bytes"
