+++ promflux/selector.go
@@ -1,4 +1,4 @@
-package main
+package promflux
 
 import (
 	"fmt"
